added gems and gem ui
created a static gamemanager script(an unmatched script that cannot perform functions but can be a project wide holder of variables) created 3 prefabs for gems and 1 prefab for a ui element to display gemcounts. created a script for each gem that updates the gamemanager corresponding counter then destroys the gem from the scene, a script on the ui element that updates the ui whenever the gems change. i also made my own player controller because i didn't know how to work the other one and just needed a simple tester for the collision with the gems